requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
attrs==18.2.0
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
idna==2.8                 # via requests
requests==2.21.0
urllib3==1.24.1           # via requests
